==> MATT/notebook/manu_version/0. Process_Data.ipynb <==
# Process Data

## 0. Preliminary

%matplotlib inline

import IPython.display as ipd
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import MultiLabelBinarizer
import os
import platform
import sys
""

sns.set_context("notebook", font_scale=1.5)
plt.rcParams['figure.figsize'] = (17, 5)

if 'Windows' in platform.platform():
    ROOT_PATH = "D:/PycharmProjects/HMAN"
else:
    ROOT_PATH = "/home/xkliu/PycharmProjects/HMAN"
RAW_DATA_PATH = ROOT_PATH  + "/raw_data"
DATA_PATH = ROOT_PATH + "/data"
os.chdir(ROOT_PATH)
sys.path.append("./")

from kddirkit.utils import utils


## 1. Load Data

tracks = utils.load(RAW_DATA_PATH+'/fma_metadata/tracks.csv')
genres = utils.load(RAW_DATA_PATH+'/fma_metadata/genres.csv')
features = utils.load(RAW_DATA_PATH+'/fma_metadata/features.csv')
echonest = utils.load(RAW_DATA_PATH+'/fma_metadata/echonest.csv')

np.testing.assert_array_equal(features.index, tracks.index)
assert echonest.index.isin(tracks.index).all()

tracks.shape, genres.shape, features.shape, echonest.shape

tracks.loc[:, ('set', 'split')] == 'training'

tracks[tracks.loc[:, ('set', 'subset')] == 'medium']

from collections import Counter

Counter(tracks.loc[:, ('set', 'subset')])

track_use = tracks.loc[:,[('album', 'id'),('album','type'),('artist','id'),
              ('set', 'split'),('set', 'subset'),('track','genre_top'),
              ('track','genres'), ('track','genres_all'),
              ('track', 'title')]]
track_use

track_use.columns

track_use.columns= ['album_id',  'album_type', 'artist_id',
                    'set_split', 'set_subset',  'track_genre_top',
                    'track_genres', 'track_genres_all', 'track_title']
track_use

## 2. Export Data

track_use.to_csv(DATA_PATH+'/tracks.csv')

track_data =  pd.read_csv(DATA_PATH+'/tracks.csv')
track_data

small_data = track_data[track_data['set_subset']=='small']
small_data_train = small_data[small_data['set_split']=='training']
small_data_test = small_data[small_data['set_split']=='test']
small_data_val = small_data[small_data['set_split']=='validation']

small_data.to_csv(RAW_DATA_PATH+'/small_data.csv', index = False)
small_data_train.to_csv(RAW_DATA_PATH+'/small_data_train.csv', index = False)
small_data_test.to_csv(RAW_DATA_PATH+'/small_data_test.csv', index = False)
small_data_val.to_csv(RAW_DATA_PATH+'/small_data_val.csv', index = False)

small_data

medium_data = track_data[track_data['set_subset']!='large']
medium_data_train = medium_data[medium_data['set_split']=='training']
medium_data_test = medium_data[medium_data['set_split']=='test']
medium_data_val = medium_data[medium_data['set_split']=='validation']

medium_data.to_csv(RAW_DATA_PATH+'/medium_data.csv', index = False)
medium_data_train.to_csv(RAW_DATA_PATH+'/medium_data_train.csv', index = False)
medium_data_test.to_csv(RAW_DATA_PATH+'/medium_data_test.csv', index = False)
medium_data_val.to_csv(RAW_DATA_PATH+'/medium_data_val.csv', index = False)

#medium_data


# medium_all_data = track_data[track_data['set_subset'] !="large"]
# medium_all_data_train = medium_all_data[medium_all_data['set_split'] == 'training']
# medium_all_data_test = medium_all_data[medium_all_data['set_split'] == 'test']
# medium_all_data_val = medium_all_data[medium_all_data['set_split'] == 'validation']
#
# medium_all_data.to_csv(RAW_DATA_PATH + '/medium_all_data.csv', index=False)
# medium_all_data_train.to_csv(RAW_DATA_PATH + '/medium_all_data_train.csv', index=False)
# medium_all_data_test.to_csv(RAW_DATA_PATH + '/medium_all_data_test.csv', index=False)
# medium_all_data_val.to_csv(RAW_DATA_PATH + '/medium_all_data_val.csv', index=False)
# medium_all_data

large_data = track_data
large_data_train = large_data[large_data['set_split']=='training']
large_data_test = large_data[large_data['set_split']=='test']
large_data_val = large_data[large_data['set_split']=='validation']

large_data.to_csv(RAW_DATA_PATH+'/large_data.csv', index = False)
large_data_train.to_csv(RAW_DATA_PATH+'/large_data_train.csv', index = False)
large_data_test.to_csv(RAW_DATA_PATH+'/large_data_test.csv', index = False)
large_data_val.to_csv(RAW_DATA_PATH+'/large_data_val.csv', index = False)
large_data

# large_all_data = track_data
# large_all_data_train = large_all_data[large_all_data['set_split'] == 'training']
# large_all_data_test = large_all_data[large_all_data['set_split'] == 'test']
# large_all_data_val = large_all_data[large_all_data['set_split'] == 'validation']
#
# large_all_data.to_csv(RAW_DATA_PATH + '/large_all_data.csv', index=False)
# large_all_data_train.to_csv(RAW_DATA_PATH + '/large_all_data_train.csv', index=False)
# large_all_data_test.to_csv(RAW_DATA_PATH + '/large_all_data_test.csv', index=False)
# large_all_data_val.to_csv(RAW_DATA_PATH + '/large_all_data_val.csv', index=False)
# large_all_data
